+++ R/db_clust_dbscan.R
@@ -1,0 +1,11 @@
+#' Density-Based Spatial Clustering of Applications with Noise (DBSCAN) via dbscan
+#'
+#' [db_clust()] creates DBSCAN model.
+#'
+#' @includeRmd man/rmd/db_clust_dbscan.md details
+#'
+#' @name details_db_clust_dbscan
+#' @keywords internal
+NULL
+
+# See inst/README-DOCS.md for a description of how these files are processed