+++ R/update.R
@@ -5,6 +5,8 @@
 #'
 #' @inheritParams k_means
 #' @inheritParams hier_clust
+#' @inheritParams db_clust
+#' @inheritParams gm_clust
 #' @param object A cluster specification.
 #' @param parameters A 1-row tibble or named list with _main_ parameters to
 #'   update. Use **either** `parameters` **or** the main arguments directly when